+++ OSBarcodeLibTests/OSBARCTestValues.swift
@@ -1,0 +1,7 @@
+struct OSBARCScannerStubValues {
+    static let scannedCode = "Scanned Code"
+}
+
+struct OSBARCCommonValues {
+    static let failMessage = "Shouldn't get here"
+}