+++ docs/CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-### 2022-04-12
+### [1.0.0]
 
-Create repository.
+- Implement iPad user interface (https://outsystemsrd.atlassian.net/browse/RMET-2911).
+- Implement iPhone user interface (https://outsystemsrd.atlassian.net/browse/RMET-2771).
+- Add Scan Orientation Selection feature (https://outsystemsrd.atlassian.net/browse/RMET-2753).
+- Add Camera Selection feature (https://outsystemsrd.atlassian.net/browse/RMET-2754).
+- Add Scan Button feature (https://outsystemsrd.atlassian.net/browse/RMET-2752).
+- Add Scan Instruction text feature (https://outsystemsrd.atlassian.net/browse/RMET-2751).
+- Add Torch Button feature (https://outsystemsrd.atlassian.net/browse/RMET-2749).
+- Implement the Scan Barcode feature (https://outsystemsrd.atlassian.net/browse/RMET-2748).